extract types to file
