+++ src/components/Inputs/TravelersRoomsInput/TravelerRoomsInputTypes.ts
@@ -1,0 +1,11 @@
+import { ButtonInputProps } from '../types';
+
+export interface TravelersRoomsInputProps extends ButtonInputProps {
+    placeholder?:ITravelersRooms;
+    value: ITravelersRooms;
+  }
+
+  interface ITravelersRooms{
+    travelers:string;
+    rooms:string;
+  }
